playwright_tests/requirements.txt
playwright>=1.40.0
pytest>=7.4.0
pytest-xdist>=3.8.0
pytest-playwright>=0.4.3
pytest-html>=4.1.0
pyyaml>=6.0
requests>=2.31.0
opencv-python-headless>=4.8.0
numpy>=1.24.0
ultralytics>=8.0.0
torch>=2.0.0
Pillow>=10.0.0
python-dotenv>=1.0.0
